+++ app/models/medication_variable.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+# Describes a variable being collected along with the auto-complete values
+class MedicationVariable < ApplicationRecord
+  # Concerns
+  include Deletable
+  include Searchable
+
+  # Validations
+  validates :name, presence: true,
+                   uniqueness: { case_sensitive: false, scope: :project_id }
+
+  # Relationships
+  belongs_to :project
+end