+++ test/models/medication_variable_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class MedicationVariableTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end