+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Medication start and stop dates allow for fuzzy dates
     - Dates where only the year is known: 1990
     - Dates where only the year and month are known: Jan 1990
+  - Project editors can configure additional variables to be collected alongside
+    medication name and start and stop dates
 
 ### Refactoring
 - Remove skipping_utf8 in search forms as Rails 6 no longer enforces utf8